changed vector data types bc im a dummy

+++ include/network.h
@@ -17,13 +17,13 @@ public:
 
   arma::mat feedforward(arma::mat a);
 
-  void stochastic_gradient_descent(std::vector<std::pair<int, std::vector<int>>>& training_data, int epochs, int mini_batch_size, float eta, std::vector<std::pair<int, std::vector<int>>> test_data = {});
+  void stochastic_gradient_descent(std::vector<std::pair<int, std::vector<double>>>& training_data, int epochs, int mini_batch_size, float eta, std::vector<std::pair<int, std::vector<double>>> test_data = {});
 
-  void update_mini_batch(std::vector<std::pair<int, std::vector<int>>> mini_batch, float eta);
+  void update_mini_batch(std::vector<std::pair<int, std::vector<double>>> mini_batch, float eta);
 
-  std::pair<std::vector<arma::mat>,std::vector<arma::mat>> backprop(std::vector<int>& x, int y);
+  std::pair<std::vector<arma::mat>,std::vector<arma::mat>> backprop(std::vector<double>& x, int y);
 
-  int evaluate(std::vector<std::pair<int, std::vector<int>>>& test_data);
+  int evaluate(std::vector<std::pair<int, std::vector<double>>>& test_data);
 
 arma::mat cost_derivative(arma::vec output_activation, int y);
